Restore remote editing
This occour when we edit remote file and plugin was reloaded. After this
SESSIONS is empty. So...
When we start remote edit iterate all views that related to 'rsub'
if it is related

    bool( env )

and is not in current session (belongs to old disconnected session)

    not v.id() in SESSIONS

and file were opened from same host

    self.env['host'] == env['host']

and same file were opened

    self.env['real-path'] == env['real-path']

So we put this view to current session in bring_to_front

                self.env['work_dir'] =  env['work_dir']
                SESSIONS[ v.id() ] =  self
                self.temp_file =  v.file_name()
                self.send_save()
                sublime.active_window().focus_view( v )
                bring_to_front()
                return